refactor(registry): split report analytics-v2 under 200-line policy
